requirements.txt
python-dotenv==1.0.1
aiogram==3.13.1
aiomysql==0.2.0